router-example/requirements.txt
python-dotenv==1.0.1
openai==1.60.0
pydantic==2.10.5
instructor==1.7.2